Rename TrustlyConfig to TrustlySettings
