+++ Sources/TrustlySDK/Settings/SettingsManager.swift
@@ -10,25 +10,25 @@ import Foundation
 
 /** @abstract Will check if sdk need to call the settings endpoint, or just return the value stored in the cache.
  @param establish: [AnyHashable : Any]
- @param completionHandler: @escaping(TrustlyConfig?) -> Void
+ @param completionHandler: @escaping(TrustlySettings?) -> Void
  */
-func getTrustlySettingsWith(establish: [AnyHashable : Any], completionHandler: @escaping(TrustlyConfig?) -> Void) {
+func getTrustlySettingsWith(establish: [AnyHashable : Any], completionHandler: @escaping(TrustlySettings?) -> Void) {
     
-    if let trustlyConfig: TrustlyConfig = readDataFrom(keyStorage: .settings),
-       trustlyConfig.isValid() {
+    if let trustlySettings: TrustlySettings = readDataFrom(keyStorage: .settings),
+       trustlySettings.isValid() {
         
-        completionHandler(trustlyConfig)
+        completionHandler(trustlySettings)
         
     } else {
-        APIRequest.getTrustlyConfigWith(establish: establish) { trustlyConfig in
+        APIRequest.getTrustlySettingsWith(establish: establish) { trustlySettings in
             
-            if let trustlyConfig = trustlyConfig {
-                let config = TrustlyConfig(settings: trustlyConfig.settings, createdDateTime: Date())
+            if let trustlySettings = trustlySettings {
+                let settings = TrustlySettings(settings: trustlySettings.settings, createdDateTime: Date())
                 
-                saveData(config, keyStorage: .settings)
+                saveData(settings, keyStorage: .settings)
             }
             
-            completionHandler(trustlyConfig)
+            completionHandler(trustlySettings)
         }
     }
 }
